refactor: extract assistant message pipeline
